+++ src/main/java/com/elmakers/mine/bukkit/plugins/magicworlds/entities/ReplaceSpawnHandler.java
@@ -1,0 +1,24 @@
+package com.elmakers.mine.bukkit.plugins.magicworlds.entities;
+
+import java.util.Map.Entry;
+
+import org.bukkit.configuration.ConfigurationSection;
+import org.bukkit.entity.EntityType;
+
+public class ReplaceSpawnHandler extends MagicSpawnHandler {
+	
+	@SuppressWarnings("deprecation")
+	public void onLoad(ConfigurationSection config) {
+		ConfigurationSection replaceList = config.getConfigurationSection("replace");
+		for (Entry<String, Object> replaceEntry : replaceList.getValues(false).entrySet()) {
+			try {
+				EntityType fromType = EntityType.fromName(replaceEntry.getKey());
+				EntityType toType = EntityType.fromName(replaceEntry.getValue().toString());
+				addRule(new SpawnReplaceRule(fromType, toType));
+				controller.getLogger().info(" Replacing: " + replaceEntry.getKey() + " with " + replaceEntry.getValue());
+			} catch (Exception ex) {
+				controller.getLogger().warning(" Invalid entity type: " + replaceEntry.getKey() + " => " + replaceEntry.getValue());
+			}
+		}
+	}
+}